+++ server/service/ParsedownExtension.php
@@ -7,6 +7,53 @@ require_once __DIR__ . "/ext/Parsedown.php";
 class ParsedownExtension extends Parsedown
 {
     /**
+     * The private attribute to access the instance of the service class
+     * UserInput.
+     */
+    private $user_input;
+
+    /**
+     * The constructor to add the new inline parser for user form fields.
+     *
+     * @param instance $user_input
+     *  An instance of the service class UserInput.
+     */
+    function __construct($user_input = null)
+    {
+        $this->user_input = $user_input;
+        $this->InlineTypes['@'][]= 'UserFormField';
+        $this->inlineMarkerList .= '@';
+    }
+
+    /**
+     * Extend the parsedown parser with user form fields. Using the syntax
+     * defined in UserInput::get_input_value_pattern() user form fields can be
+     * placed in markdown and, thus, allow to format user input fileds in any
+     * way imaginable (as HTML is allowed in markdown).
+     */
+    protected function inlineUserFormField($excerpt)
+    {
+        if($this->user_input === null)
+            return null;
+        $pattern = '/' . $this->user_input->get_input_value_pattern() . '/';
+        if (preg_match($pattern, $excerpt['text'], $matches))
+        {
+            $value = $this->user_input->get_input_value_by_pattern($matches[0],
+                $_SESSION['id_user']);
+            return array(
+
+                // How many characters to advance the Parsedown's
+                // cursor after being done processing this tag.
+                'extent' => strlen($matches[0]),
+                'element' => array(
+                    'name' => 'span',
+                    'text' => $value ?? "Bad Syntax",
+                ),
+            );
+        }
+    }
+
+    /**
      * Extende the inlineImage parser by changing the base path of image
      * sources. Further, allow to specify the image width and height as well
      * as css classes that will be attached to the image. By default the classes
@@ -17,7 +64,7 @@ class ParsedownExtension extends Parsedown
      *
      * The size and class attributes are postfixe to the filename, delimited by
      * a colon:
-     * \<image_src\>:\<width\>x\<height\>:\<class_n\>,...,\<class_n\>
+     * \<image_src\>|\<width\>x\<height\>|\<class_n\>,...,\<class_n\>
      */
     protected function inlineImage($excerpt)
     {
@@ -28,7 +75,7 @@ class ParsedownExtension extends Parsedown
             return null;
         }
 
-        $attrs = explode(':', $image['element']['attributes']['src']);
+        $attrs = explode('|', $image['element']['attributes']['src']);
         $image['element']['attributes']['src'] = $attrs[0];
         if(count($attrs) > 1)
         {
@@ -46,11 +93,36 @@ class ParsedownExtension extends Parsedown
                 "img-fluid img-thumbnail " . $class;
         }
 
-        if($image['element']['attributes']['src'][0] !== '/')
-            $image['element']['attributes']['src'] = ASSET_PATH . '/'
-                . $image['element']['attributes']['src'];
-
         return $image;
+    }
+
+    /**
+     * Extends the inlineLink parser to add the following features:
+     *
+     * - prepending a link url with '!' will open the link in a new tab
+     * - prepending a link url with '%' will prepend the asset path
+     * - prepending a link url with '|' will prepend the base path
+     */
+    protected function inlineLink($excerpt)
+    {
+        $link = parent::inlineLink($excerpt);
+
+        if($link['element']['attributes']['href'][0] === '!')
+        {
+            $link['element']['attributes']['target'] = '_blank';
+            $link['element']['attributes']['href'] =
+                substr($link['element']['attributes']['href'], 1);
+        }
+        if($link['element']['attributes']['href'][0] === '%')
+        {
+            $link['element']['attributes']['href'] = ASSET_PATH . '/'
+                . substr($link['element']['attributes']['href'], 1);
+        }
+        else if($link['element']['attributes']['href'][0] === '|')
+            $link['element']['attributes']['href'] = BASE_PATH . '/'
+                . substr($link['element']['attributes']['href'], 1);
+
+        return $link;
     }
 
     /**