+++ CHANGELOG.md
@@ -12,6 +12,7 @@
  - Change CSP rule to allow images from https://via.placeholder.com/
  - Update Bootstrap to version v4.3.1
  - Distinguish between export and experiment activity
+ - Change the image syntax in markdown to allow more linking options
 
 ### New Features
 
@@ -24,12 +25,15 @@
    (#80)
  - Allow to customize the `meta:description` tag (#164)
  - Add a `mermaidForm` style which allows to describe graphs and allows a
-   subject to change node labels (!2)
+   subject to change node labels (#162)
  - Improve the CMS:
    - add a schematic page overview
    - allow to jump to the real page
    - if acl allows it, a user sees a small edit icon on to bottom right of each
      section page which allows an experimenter to switch back to the CMS.
+ - Improve Markdown
+   - Allow more linking options (use prefixes '|', '!', '%')
+   - Allow to fetch user form input fields (#154)
 
 ------------
 # latest